integrations/CrewAI-LangGraph/requirements.txt
crewai==0.130.0
langgraph==0.0.15
langchain-community==0.3.26
python-dotenv==1.0.0
google-search-results==2.1.0
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
tavily-python==0.3.1
